2.2.4 Renamed package ('list' to 'stack')
